Add brokers as follower and leader when decommissioning brokers.
The best of the two will be picked.

When replacing a broker with a new one, it often happens that the new
broker is empty and can be picked as leader for all/most of the
partitions on the broker to be decommissioned.
This change will explore adding each broker to the replica set of
partitions as a follower, and as a leader.