Remove default value of float(inf) for --max-movement-size in parser
Having this value in the parser means that if the user does not pass in
any value, it'll still use the float(inf) value in the genetic balancer
for decommission. We want to make this more similar to how rebalance
works, so removing default value in the parser makes sense.