Allow using Kafka AdminClient for kafka-consumer-manager commands
Reading the __consumer_offsets topic at every run is slow and can add
unwanted load on a Kafka cluster. The AdminClient allows to get the list
of consumer groups and the list of topics from a group id in constance
time.

This is currently a feature flag to keep backwards compatibility, but
will later be enabled by default.

--use-admin-client